(#8975) Add minimum_versionable package feature
Issues 8975 and 12179 both request that an additional means of
specifying package version(s) be provided, specifically the ability to
upgrade a version when it is less than a specified minimum. This commit
introduces a new feature to the package provider called
'minimum_versionable', which will, with provider support, selectively
install the latest available package dependent on whether or not the
currently installed package is less than the supplied version number.