Fix reverse ordering of node classes
When compiling a catalog for a node the compiler first partitions
parameterized classes from non-parameterized classes, and then applies
each set individually.

In the original implementation, for parameterized classes, the classes
would be applied in reverse-order when compared to the data given by the
node object. Furthermore, the classes without parameters specified would
be evaluated first, followed by the classes that did specify parameters.

This behavior caused two problems.

First, when dealing with parameterized classes Puppet is actually
evaluation-order dependent. Therefore it is desirable, in order to
preserve user intent, to evaluate classes in the order received.
Reverse-sorting the list of classes effectively guaranteed that the user
would not be able to sanely pass ordered lists of classes.

Second, when specifying classes without parameters, it is possible that
during evaluation of those non-parameterized classes one of the
parameterized classes will be included or inherited.

If the parameterized classes were evaluated first, this would not cause
any problems. But as the classes without parameters are evaluated first,
this could introduce avoidable duplicate class declaration problems.